// src/app/core/services/model/reportServiceParametersDto.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { EventTypes } from './eventTypes';
import { ReportDataQueryParameters } from './reportDataQueryParameters';
import { SftpOptionsInput } from './sftpOptionsInput';

export interface ReportServiceParametersDto { 
    /**
     * id of the service (rowId)
     */
    serviceId?: number;
    /**
     * user uuid of the service creator
     */
    createdByUuid?: string;
    /**
     * name of the service
     */
    serviceName?: string;
    /**
     * name of the files created by the service
     */
    fileNameWithoutExtension?: string;
    /**
     * what should the service do, if it gets triggered?  possible options: write usbfile, upload file to sftp-server, send email
     */
    action?: string;
    /**
     * if Action = email, send the service output to this email address
     */
    emailAddress?: string;
    sftpOptionsInput?: SftpOptionsInput;
    /**
     * how gets the service triggered?  possible options: event (e.g. new ALARM), manual, cronjob, each x records (data log entries)
     */
    trigger?: string;
    /**
     * if Trigger = event, defines the trigger events, e.g. TriggerMessageTypes = \"ALARM | DISRUPTION\"
     */
    triggerMessageTypes?: Array<EventTypes>;
    /**
     * if Trigger = data, we send a report after each TriggerDataIntervalCount new entries in data_dataLog
     */
    triggerDataIntervalCount?: number;
    /**
     * if Trigger = cronjob, defines the cronjob execution time
     */
    cronJobTimeStr?: string;
    dataQueryParameters?: ReportDataQueryParameters;
}